break out the watchers method from helper to avoid the recursive require
