+++ src/lib/utils/helper.js
@@ -8,8 +8,8 @@ const path = require('path');
 const _ = require('lodash');
 const express = require('express');
 // Our modules
-const logutil = require('./log.js');
-const streamWatcher = require('./stream-watcher');
+const logutil = require('./log');
+// const streamWatcher = require('./stream-watcher');
 /**
  * create a random number between two values, for creating a random port number
  * @param {int} min
@@ -120,41 +120,7 @@ const proxyDelay = (originalUrl, delayReqTime, delayResTime) => {
   }
 };
 
-/**
- * Watcher - moving back from the gulp.js export
- * Rename from watcher --> fileWatcher 
- * @param {mixed} filePaths array of string 
- * @param {function} callback function to execute when file change 
- * @param {boolean} verbose param pass to the streamWatcher should show console.log or not
- * @param {int} debounce ms to determine when the callback should execute
- * @return {function} the streamWatcher terminate callback
- */
-const fileWatcher = (filePaths, callback, verbose = true, debounce = 300) => {
-  let files = [];
-  return streamWatcher(filePaths, verbose)
-    .doAction(f => files.push(f))
-    .debounce(debounce)
-    .onValue(() => {
-      if (files.length) {
-        callback(files);
-        // Reset
-        files = [];
-      }
-    });
-};
 
-/**
- * Wrapper for the serverReload option - this will run in it's own process
- * @TODO try to figure out how to run this in a different process to avoid too many watchers
- * @param {object} config we pass the options.serverReload here 
- * @return {mixed} ps config.enable or false
- */
-const serverReload = config => {
-  if (config.enable && _.isFunction(config.callback)) {
-    return fileWatcher(config.dir, config.callback, config.config.verbose, config.config.debounce);
-  }
-  return () => {};
-};
 
 /**
  * directory listing - no longer support since 1.4.0-alpha.2
@@ -171,7 +137,5 @@ module.exports = {
   getRandomInt: getRandomInt,
   toArray: toArray,
   proxyDelay: proxyDelay,
-  serveStatic: serveStatic,
-  fileWatcher: fileWatcher,
-  serverReload: serverReload
+  serveStatic: serveStatic
 };
